+++ Subzy/MauiProgram.cs
@@ -31,12 +31,11 @@ public static class MauiProgram
 		builder.Services.AddSingleton<IImageProcessor, ImageProcessorService>();
 		builder.Services.AddSingleton<ChangeDetectorService>();
 #if ANDROID
-		builder.Services.AddSingleton<IOcrService, Platforms.Android.Services.MlKitOcrService>();
+        builder.Services.AddSingleton<IOcrService, Platforms.Android.Services.TesseractOcrService>();
 #else
 		// No OCR service available for other platforms
-		builder.Services.AddSingleton<IOcrService, TesseractOcrService>();
 #endif
-		builder.Services.AddSingleton<ITranslationService, AzureTranslatorService>();
+        builder.Services.AddSingleton<ITranslationService, AzureTranslatorService>();
 		builder.Services.AddSingleton<ITtsService, AzureTtsService>();
 		builder.Services.AddSingleton<ForegroundAppDetector>();
 		builder.Services.AddSingleton<ColorProfileManager>();